+++ src/commands/minion.ts
@@ -3,12 +3,16 @@ import { chunk } from 'e';
 import { CommandStore, KlasaMessage } from 'klasa';
 
 import { ClueTiers } from '../lib/clues/clueTiers';
-import { Emoji, lastTripCache } from '../lib/constants';
+import { Emoji, lastTripCache, PerkTier } from '../lib/constants';
 import { requiresMinion } from '../lib/minions/decorators';
 import { runCommand } from '../lib/settings/settings';
 import { BotCommand } from '../lib/structures/BotCommand';
+import { getUsersTame, shortTameTripDesc, tameLastFinishedActivity } from '../lib/tames';
 import { convertMahojiResponseToDJSResponse } from '../lib/util';
+import getUsersPerkTier from '../lib/util/getUsersPerkTier';
 import { makeDoClueButton, makeRepeatTripButton } from '../lib/util/globalInteractions';
+import { getItemContractDetails } from '../mahoji/commands/ic';
+import { spawnLampIsReady } from '../mahoji/commands/tools';
 import { calculateBirdhouseDetails } from '../mahoji/lib/abstracted_commands/birdhousesCommand';
 import { isUsersDailyReady } from '../mahoji/lib/abstracted_commands/dailyCommand';
 import { canRunAutoContract } from '../mahoji/lib/abstracted_commands/farmingContractCommand';
@@ -96,12 +100,51 @@ export default class MinionCommand extends BotCommand {
 			extraButtons.push(makeRepeatTripButton());
 		}
 
+		const mahojiUser = await mahojiUsersSettingsFetch(msg.author.id);
+		const [spawnLampReady] = spawnLampIsReady(mahojiUser, msg.channel.id);
+		if (spawnLampReady) {
+			extraButtons.push(
+				new MessageButton()
+					.setLabel('Spawn Lamp')
+					.setEmoji('988325171498721290')
+					.setCustomID('SPAWN_LAMP')
+					.setStyle('SECONDARY')
+			);
+		}
+
+		const icDetails = getItemContractDetails(mahojiUser);
+		if (msg.author.perkTier >= PerkTier.Two && icDetails.currentItem && icDetails.owns) {
+			extraButtons.push(
+				new MessageButton()
+					.setLabel(`IC: ${icDetails.currentItem.name.slice(0, 20)}`)
+					.setEmoji('988422348434718812')
+					.setCustomID('ITEM_CONTRACT_SEND')
+					.setStyle('SECONDARY')
+			);
+		}
+
 		const bank = msg.author.bank();
 
 		for (const tier of ClueTiers.filter(t => bank.has(t.scrollID))
 			.reverse()
 			.slice(0, 3)) {
 			extraButtons.push(makeDoClueButton(tier));
+		}
+
+		if (getUsersPerkTier(msg.author) >= PerkTier.Two) {
+			const { tame, species, activity } = await getUsersTame(msg.author);
+			if (tame && !activity) {
+				const lastTameAct = await tameLastFinishedActivity(mahojiUser);
+				if (lastTameAct) {
+					extraButtons.push(
+						new MessageButton()
+							.setLabel(`Repeat ${shortTameTripDesc(lastTameAct)}`)
+							.setEmoji(species!.emojiID)
+							.setCustomID('REPEAT_TAME_TRIP')
+							.setStyle('SECONDARY')
+					);
+				}
+			}
 		}
 
 		return msg.channel.send({ content: msg.author.minionStatus, components: chunk(extraButtons, 5) });
@@ -135,7 +178,7 @@ export default class MinionCommand extends BotCommand {
 	async buy(msg: KlasaMessage) {
 		return msg.channel.send(
 			convertMahojiResponseToDJSResponse(
-				await minionBuyCommand(await mahojiUsersSettingsFetch(msg.author.id), false)
+				await minionBuyCommand(msg.author, await mahojiUsersSettingsFetch(msg.author.id), false)
 			)
 		);
 	}
