+++ src/lib/leagues/leagues.ts
@@ -17,7 +17,6 @@ import {
 	personalWoodcuttingStats
 } from '../../mahoji/lib/abstracted_commands/statCommand';
 import { getSkillsOfMahojiUser, getUserGear, mahojiUsersSettingsFetch } from '../../mahoji/mahojiSettings';
-import { ClueTiers } from '../clues/clueTiers';
 import { getParsedStashUnits } from '../clues/stashUnits';
 import { calcCLDetails } from '../data/Collections';
 import { roboChimpUserFetch } from '../roboChimp';
@@ -34,6 +33,7 @@ import { hardTasks } from './hardTasks';
 import { betterHerbloreStats, HasFunctionArgs, Task } from './leaguesUtils';
 import { masterTasks } from './masterTasks';
 import { mediumTasks } from './mediumTasks';
+import { calcActualClues } from './stats';
 
 export const leagueTasks = [
 	{ name: 'Easy', tasks: easyTasks, points: 20 },
@@ -109,38 +109,6 @@ function calcSuppliesUsedForSmithing(itemsSmithed: Bank) {
 		input.add(new Bank(smithable.inputBars).multiply(qty));
 	}
 	return input;
-}
-
-export async function calcActualClues(user: User) {
-	const result: { id: number; qty: number }[] =
-		await prisma.$queryRawUnsafe(`SELECT (data->>'clueID')::int AS id, SUM((data->>'quantity')::int) AS qty
-FROM activity
-WHERE type = 'ClueCompletion'
-AND user_id = '${user.id}'::bigint
-AND data->>'clueID' IS NOT NULL
-AND completed = true
-GROUP BY data->>'clueID';`);
-	const casketsCompleted = new Bank();
-	for (const res of result) {
-		const item = getItem(res.id);
-		if (!item) continue;
-		casketsCompleted.add(item.id, res.qty);
-	}
-	const cl = new Bank(user.collectionLogBank as ItemBank);
-	const opens = new Bank(user.openable_scores as ItemBank);
-
-	// Actual clues are only ones that you have: received in your cl, completed in trips, and opened.
-	const actualClues = new Bank();
-
-	for (const [item, qtyCompleted] of casketsCompleted.items()) {
-		const clueTier = ClueTiers.find(i => i.id === item.id)!;
-		actualClues.add(
-			clueTier.scrollID,
-			Math.min(qtyCompleted, cl.amount(clueTier.scrollID), opens.amount(clueTier.id))
-		);
-	}
-
-	return actualClues;
 }
 
 export async function calcLeaguesRanking(user: RoboChimpUser) {
