+++ src/mahoji/lib/abstracted_commands/statCommand.ts
@@ -12,7 +12,7 @@ import { ClueTiers } from '../../../lib/clues/clueTiers';
 import { getClueScoresFromOpenables } from '../../../lib/clues/clueUtils';
 import { Emoji, PerkTier } from '../../../lib/constants';
 import { calcCLDetails } from '../../../lib/data/Collections';
-import { calcActualClues } from '../../../lib/leagues/leagues';
+import { calcActualClues } from '../../../lib/leagues/stats';
 import backgroundImages from '../../../lib/minions/data/bankBackgrounds';
 import killableMonsters from '../../../lib/minions/data/killableMonsters';
 import { getMinigameScore } from '../../../lib/settings/minigames';
